test(tooling): move Rust tests beside modules
